requirements.txt
# Core packages for data handling and API interaction
google-api-python-client
youtube-transcript-api
pandas
python-dotenv
matplotlib
wordcloud

# Configuration and validation
pyyaml
pydantic

# Data analysis and visualization
scikit-learn
seaborn

# Future AI/RAG Packages
langchain-google-genai
chromadb
google-genai